Add files via upload
 web scraper is an automated tool that can extract large amounts of data from sites. You can collect up to hundreds of thousands of data points in just a few minutes with the help of web scraping.

We will scrape a site called “Quotes to Scrape” in this tutorial. This is a site that was specifically built to practice web scraping.

By the end of this article, you will be familiar with:

Extracting raw HTML from a website
Using the BeautifulSoup library to parse this HTML and extract useful pieces of information from the site
Collecting data from multiple webpages at once
Storing this data into a Pandas dataframe.